+++ conf_engine/parsers/__init__.py
@@ -1,5 +1,5 @@
-import config_engine.parsers.env
-import config_engine.parsers.ini_file
+import conf_engine.parsers.env
+import conf_engine.parsers.ini_file
 
 EnvironmentParser = env.EnvironmentParser
 INIFileParser = ini_file.INIFileParser