+++ conf_engine/tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-import config_engine.configuration
+import conf_engine.configuration
 
 
 @pytest.fixture
@@ -8,7 +8,7 @@ def test_ini_directory(request):
     path = ''
     for part in request.path.parts[1:]:
         path += '/' + part
-        if part == 'config_engine':
+        if part == 'conf_engine':
             return path + '/tests/test_ini_files'
 
 
@@ -19,4 +19,4 @@ def test_config(monkeypatch, cli_opts: [str] = None, ):
         cli_opts.insert(0, 'program')
 
     monkeypatch.setattr('sys.argv', cli_opts)
-    return config_engine.configuration.Configuration()
+    return conf_engine.configuration.Configuration()