+++ docs/src/examples/groups.py
@@ -1,5 +1,5 @@
-from config_engine import config
-from config_engine import options
+from conf_engine import config
+from conf_engine import options
 
 # Create option list.
 api1_options = [