+++ CHANGELOG.rst
@@ -1,10 +1,15 @@
 Changelog
 ---------
 
-**[0.4]**
+**[1.0.1]**
+   - Fixed bug where options with a default of None were not returning None
+     when the option was unset in the configuration sources.
+
+**[1.0.0]**
    - Fixed bug in evaluation of precedence when an option is defined in
      more than one source.
    - Fixed bug in defaults where a default of `False` would not be returned.
+   - Convert to MIT license.
 
 **[0.3]**
    - Added `default` argument to Option.  Now Options can be declared with